+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/BandsMembersDto.kt
@@ -3,8 +3,8 @@ package org.aleks616.shrendar.band.model
 import java.io.Serializable
 
 data class BandsMembersDto(
-    val id:Int?=null,
-    val artistId:Int?=null,
+    val id:Long?=null,
+    val artistId:Long?=null,
     val artistName:String?=null,
     val bandId:Int?=null,
     val bandName:String?=null,
